ConfigMapAfter refactor
